text-minimessage: Move resolvers into their own package
