+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/tag/Placeholder.java
@@ -26,6 +26,7 @@ package net.kyori.adventure.text.minimessage.tag;
 import java.util.Objects;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.ComponentLike;
+import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
 import org.jetbrains.annotations.NotNull;
 
 import static java.util.Objects.requireNonNull;
