+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/tag/resolver/package-info.java
@@ -21,27 +21,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.tag;
-
-import java.util.Map;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-final class MapTagResolver implements TagResolver.WithoutArguments, MappableResolver {
-  private final Map<String, ? extends Tag> tagMap;
-
-  MapTagResolver(final @NotNull Map<String, ? extends Tag> placeholderMap) {
-    this.tagMap = placeholderMap;
-  }
-
-  @Override
-  public @Nullable Tag resolve(final @NotNull String name) {
-    return this.tagMap.get(name);
-  }
-
-  @Override
-  public boolean contributeToMap(final Map<String, Tag> map) {
-    map.putAll(this.tagMap);
-    return true;
-  }
-}
+/**
+ * Tools for resolving tags based on tag information.
+ *
+ * @since 4.10.0
+ */
+package net.kyori.adventure.text.minimessage.tag.resolver;
