+++ text-minimessage/src/test/java/net/kyori/adventure/text/minimessage/TestBase.java
@@ -25,7 +25,7 @@ package net.kyori.adventure.text.minimessage;
 
 import java.util.stream.Collectors;
 import net.kyori.adventure.text.Component;
-import net.kyori.adventure.text.minimessage.tag.TagResolver;
+import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
 import net.kyori.examination.string.MultiLineStringExaminer;
 import org.jetbrains.annotations.NotNull;
